+++ Type/ValueTypeInterface.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Tdn\PhpTypes\Type;
+
+/**
+ * Interface ValueInterface.
+ */
+interface ValueTypeInterface
+{
+    /**
+     * @return mixed
+     */
+    public function get();
+}